update row counting/limiting to allow values exceeding 2^32 and adjust corresponding limits in cql SELECT requests
Currently, we cannot select more than 2^32 rows from a table, because when counting the number of rows
we are using 4-byte integers as variables holding the numbers and std::numeric_limits<uint32_t>::max()
as the maximum number of rows. This patch changes these types to 8-byte integers. In order to be able to
process queries originating from older nodes, the top 32 bits of new integers are optional and are
passed a the end of a serialized class - if they're not present we assume 0.

Fixes #5101.