added a feature flag LARGE_PAGING_STATE
The flag prevents a scenario, where one node has been upgraded to a version with paging_state
holding the number of remaining rows in 64 bits, it can generate a result with such
paging_state when queried. In case we lose access to this node, we could try to fetch the
remaining pages from another node. If the node hasn't been upgraded yet, it would  only look
at the bottom 32 bits of the numer of remaining rows and return possibly much less results
than requested i.e. when the number slightly exceeds 2^32.

Until the whole cluster is upgraded, the maximum number of rows in a result it 2^32.