sstables: fold row_consumer into mp_row_consumer_k_l
The split between row_consumer and mp_row_consumer_k_l made sense for the original sstable,
which was just a sorted strings table,  and so a lower layer that doesn't understand the
schema is possible. But it doesn't make sense now.
The closely related class data_consume_rows_context now has a reference to mp_row_consumer_k_l
instead of row_consumer, so the class definition is moved to the former place of row_consumer.
The class row_consumer was used in some 2015 tests, which have been now retired.

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>