sstables: push reversing reader down to sstable::make_reader_v1
Before we implement efficient reversed sstable reads, in order to
prepare the layers above for a seamless switch, we move the
`make_reversing_reader` invocation on sstable reader from `table::make_reader`
into `sstable::make_reader_v1`; the API of the latter function is now
the same as it will be when the efficient reader is implemented (modulo
forwarding which won't be supported in the initial version).

Unforunately we don't have a `make_reversing_reader` for the v2 format,
so `sstable::make_reader`, which returns a `flat_mutation_reader_v2`,
does not support reversed queries at this point (an appropriate comment
was added). Fortunately nothing uses it in such context yet. We will
lift this restriction partially after we implement reversing directly in
the mx reader for single partition reads.