memtable: store unreversed slice for reverse multi-partition queries
This commit is a preparation for later commits, which will modify the
API of sstable reader: right now, when the sstable reader is given a
reversed slice, it performs a non-reversed query (reversing is expected
to be performed by the caller). The new API will have the sstable reader
itself perform the reversing.

This would introduce an incompatibility
with the memtable reader, which in certain scenarios delegates the work
to an underlying sstable reader (specifically: when there was a flush
during a multi-partition query). Indeed: the memtable will still expect
the results to be returned in forward order (we do not implement
reversed queries in the memtable reader yet), but the sstable reader
would already return them in reverse order according to the new API.

We fix that by ensuring that the slice passed to the underlying sstable
reader is always non-reversed. From an implementation point of view,
this commit is a no-op: right now every reader which obtains a reversed
slice would unreverse it anyway by using
`query::clustering_key_filter_ranges::get_ranges` (which does the
unreversing). We simply do it a bit earlier.

The fix will have to be removed when reversing is directly implemented
in the memtable reader, so the memtable reader will expect reversed
results from the underlying reader during reversed queries.

Note: multi-partition reversed queries are only available from Thrift at
this point and probably nobody uses this, so all this work probably does
not matter.